refact: clean unnecessary files
