+++ SNUTT-2022/SNUTT/AppState/AppState.swift
@@ -10,7 +10,6 @@ import SwiftUI
 
 class AppState {
     var user = UserState()
-    var setting = Setting()
 
     var system = SystemState()
     var search = SearchState()
