--- SNUTT-2022/SNUTT/AppState/States/Setting.swift
@@ -1,14 +1,0 @@
-//
-//  Setting.swift
-//  SNUTT
-//
-//  Created by Jinsup Keum on 2022/06/25.
-//
-
-import Combine
-import SwiftUI
-
-class Setting: ObservableObject {
-    var colorList: String? // STColorList
-    var snuevWebUrl: String?
-}
