--- SNUTT-2022/SNUTT/Repositories/Dto/CourseBook.swift
@@ -1,8 +1,0 @@
-//
-//  CourseBook.swift
-//  SNUTT
-//
-//  Created by 박신홍 on 2022/08/19.
-//
-
-import Foundation
