--- SNUTT-2022/SNUTT/Repositories/Dto/SearchResultDto.swift
@@ -1,8 +1,0 @@
-//
-//  SearchResultDto.swift
-//  SNUTT
-//
-//  Created by 박신홍 on 2022/07/21.
-//
-
-import Foundation
