+++ SNUTT-2022/SNUTT/Views/Components/WebViews/LicenseView.swift
