+++ SNUTT-2022/SNUTT/Views/Scenes/TimetableScene.swift
@@ -88,31 +88,6 @@ struct TimetableScene: View {
     }
 }
 
-// TODO: Move this to Modifiers folder.
-struct ViewDidLoadModifier: ViewModifier {
-    @State private var didLoad = false
-    private let action: (() async -> Void)?
-
-    init(perform action: (() async -> Void)? = nil) {
-        self.action = action
-    }
-
-    func body(content: Content) -> some View {
-        content.task {
-            if didLoad == false {
-                didLoad = true
-                await action?()
-            }
-        }
-    }
-}
-
-extension View {
-    /// Adds an (asynchronous) action to perform when this view is loaded.
-    func onLoad(perform action: (() async -> Void)? = nil) -> some View {
-        modifier(ViewDidLoadModifier(perform: action))
-    }
-}
 
 // struct MyTimetableScene_Previews: PreviewProvider {
 //    static var previews: some View {
